requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.23.0
psycopg2==2.9.6
pycryptodome==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
SQLAlchemy==2.0.13
toolz==0.12.0
typing_extensions==4.5.0
urllib3==2.0.2
web3==6.3.0
websockets==11.0.3
yarl==1.9.2
